Rearranging course methods.
