+++ myapp/Classes/supervisor.py
@@ -1,5 +1,6 @@
 from myapp.models import User, Course
 from myapp.Classes.users import Users
+from myapp.Classes.courses import CourseUtility
 
 
 class Supervisor(Users):
@@ -50,6 +51,20 @@ class Supervisor(Users):
             course = Course.objects.get(Course_Code=course_code)
             course.Course_Instructor=instructor_name
             course.save()
+
+    @staticmethod
+    def create_course(code, name, desc, inst):
+        courses = CourseUtility.get_course_list()
+        for course in courses:
+            if course.Course_Code == code:
+                return TypeError("A course with that code already exists!")
+        if code is "":
+            return TypeError("Course code cannot be blank!")
+        elif name is "":
+            return TypeError("Course name cannot be blank!")
+        elif desc is "":
+            return TypeError("Course description cannot be blank!")
+        Course.objects.create(Course_Code=code, Course_Name=name, Course_Description=desc, Course_Instructor=inst)
 
     @staticmethod
     def editCourse(course_name='default', course_desc='default', isonline='default', location='default',
